httpTemplater/app/requirements.txt
aiohttp
pyyaml
aioredis